:recycle: refactor(api) remove `pre-response-utils.js` from `lib/`
